+++ docs/CHANGELOG.md
@@ -29,6 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Flaw-to-Upstream Mapping Endpoints (OSIDB-5085)
 - Add Upstream Maintainer Send Email Action (OSIDB-5088)
 - Add report payloads in milestones (OSIDB-5070)
+- Add Observability for Upstream Maintainer Email Actions (OSIDB-5104)
+- Add Upstream Maintainer Notifications Query (OSIDB-5103)
 
 ### Changed
 - align existing workflows with workflow framework concepts